2_syntetic_data/pyproject.toml
[project]
name = "2-syntetic-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.59.0",
    "pydantic>=2.10.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
]
